streamlit_modern_dashboard/requirements.txt
streamlit==1.37.1
scikit-learn==1.2.2
pandas==2.0.3
numpy==1.24.4
joblib==1.3.2
pyarrow==12.0.1
pillow==10.1.0
plotly==5.18.0
xgboost==1.7.6